fix: Separate build, runtime, and test dependencies
Closes #229
Closes #239
