+++ .appveyor.yml
@@ -13,10 +13,12 @@ install:
   - conda info -a
   - conda create -q -n test-env python=%PYTHON_VERSION%
   - activate test-env
-  - conda install numpy cython scipy matplotlib pandas setuptools libpython m2w64-toolchain -c msys2 -q
+  - conda install libpython m2w64-toolchain -c msys2 -q
   - gcc --version
-  - mingw32-make --version
   # Install the build dependencies of the project.
+  - pip install -q -r build-requirements.txt
+  - mingw32-make --version
+  # Install the other dependencies of the project.
   - pip install -q -r requirements.txt
   - pip install -q -r test-requirements.txt
