+++ .travis.yml
@@ -16,7 +16,8 @@ before_install:
   - python --version
   - if [[ -n $CC ]]; then $CC --version; fi
 install:
-  - pip install cython tox coveralls grpcio-tools mypy-protobuf
+  - pip install tox
+  - pip install -r build-requirements.txt
   - make  # generate required C++ files
 script:
   - travis_wait 60 tox
