--- doc/source/httpstan.rst
@@ -1,109 +1,0 @@
-httpstan package
-================
-
-Subpackages
------------
-
-.. toctree::
-
-    httpstan.services
-
-Submodules
-----------
-
-httpstan.cache module
----------------------
-
-.. automodule:: httpstan.cache
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.callbacks\_writer\_parser module
------------------------------------------
-
-.. automodule:: httpstan.callbacks_writer_parser
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.callbacks\_writer\_pb2 module
---------------------------------------
-
-.. automodule:: httpstan.callbacks_writer_pb2
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.compile module
------------------------
-
-.. automodule:: httpstan.compile
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.fits module
---------------------
-
-.. automodule:: httpstan.fits
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.main module
---------------------
-
-.. automodule:: httpstan.main
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.models module
-----------------------
-
-.. automodule:: httpstan.models
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.schemas module
------------------------
-
-.. automodule:: httpstan.schemas
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.services\_stub module
-------------------------------
-
-.. automodule:: httpstan.services_stub
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.stan module
---------------------
-
-.. automodule:: httpstan.stan
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-httpstan.utils module
----------------------
-
-.. automodule:: httpstan.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: httpstan
-    :members:
-    :undoc-members:
-    :show-inheritance:
