--- doc/source/httpstan.services.rst
@@ -1,22 +1,0 @@
-httpstan.services package
-=========================
-
-Submodules
-----------
-
-httpstan.services.arguments module
-----------------------------------
-
-.. automodule:: httpstan.services.arguments
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: httpstan.services
-    :members:
-    :undoc-members:
-    :show-inheritance:
