+++ requirements.txt
@@ -9,7 +9,6 @@ webargs~=5.5
 marshmallow~=3.2
 numpy~=1.7
 Cython~=0.29
-apispec[yaml,validation]~=3.1
 protobuf~=3.3
 setuptools>=28.8
 ujson~=1.3
