+++ scripts/build_docs.sh
@@ -5,7 +5,7 @@ IFS=$'\n\t'
 
 OPENAPI_FILENAME=doc/source/openapi.yaml
 echo "writing OpenAPI spec to $OPENAPI_FILENAME"
-python3 -c'from httpstan import routes; print(routes.openapi_spec().to_yaml())' > "$OPENAPI_FILENAME"
+python3 -c'from httpstan import openapi; print(openapi.openapi_spec().to_yaml())' > "$OPENAPI_FILENAME"
 
 echo "Generating package API documentation with sphinx-apidoc"
 # echo excluding ``httpstan/views.py`` as Sphinx cannot process the the OpenAPI YAML
