+++ setup.py
@@ -6,5 +6,5 @@ import setuptools
 
 
 setuptools.setup(
-    setup_requires=['pbr>=2.0'],
+    setup_requires=['pbr~=5.4'],
     pbr=True)
