+++ tests/test_openapi_spec.py
@@ -1,10 +1,10 @@
 """Test OpenAPI spec generation."""
-import httpstan.routes
+import httpstan.openapi
 
 
 def test_openapi_spec() -> None:
     """Test OpenAPI spec generation."""
-    spec = httpstan.routes.openapi_spec()
+    spec = httpstan.openapi.openapi_spec()
     assert spec
     models_endpoint = spec.to_dict()["paths"]["/v1/models"]
     expected = {"$ref": "#/definitions/CreateModelRequest"}
