+++ tox.ini
@@ -14,21 +14,18 @@ deps = -r{toxinidir}/test-requirements.txt
 commands = python -m pytest -s -v tests {posargs} # substitute with tox' positional arguments
 
 [testenv:flake8]
-skip_install = true
 commands = flake8 {posargs}
 
 [testenv:format]
-skip_install = true
 deps =
     black==19.10b0
 commands = black --check httpstan tests scripts
 
 [testenv:docs]
-skip_install = true
+deps = -r{toxinidir}/build-requirements.txt
 commands = {toxinidir}/scripts/build_docs.sh
 
 [testenv:mypy]
-skip_install = true
 commands = mypy httpstan tests scripts
 
 [testenv:coverage]
